+++ common/src/main/java/smartin/miapi/config/MiapiClientConfig.java
@@ -5,6 +5,8 @@ import com.redpxnda.nucleus.codec.auto.ConfigAutoCodec;
 import com.redpxnda.nucleus.util.Color;
 import com.redpxnda.nucleus.util.Comment;
 
+import java.util.List;
+
 @ConfigAutoCodec.ConfigClassMarker
 public class MiapiClientConfig {
     @AutoCodec.Name("gui_colors")
@@ -19,6 +21,7 @@ public class MiapiClientConfig {
 
         @Comment("The color Miapi uses for its green/valid/positive color in the workbench gui")
         public Color green = new Color(0, 255, 0, 255);
+
     }
 
     @ConfigAutoCodec.ConfigClassMarker
@@ -26,5 +29,12 @@ public class MiapiClientConfig {
         @Comment("Whether Miapi materials can be animated")
         @AutoCodec.Name("animated_materials")
         public boolean animatedMaterials = true;
+        @Comment("The color Miapi uses for its enchanting glint")
+        @AutoCodec.Name("enchanting_glint_colors")
+        public List<Color> enchantColors = List.of(Color.MAGENTA);
+
+        @Comment("Speed of Color Change on enchanting Glint")
+        @AutoCodec.Name("enchanting_glint_speed")
+        public float enchantingGlintSpeed = 1.0f;
     }
 }