+++ common/src/main/java/smartin/miapi/modules/properties/potion/OnDamagedEffects.java
@@ -1,6 +1,7 @@
 package smartin.miapi.modules.properties.potion;
 
 import dev.architectury.event.EventResult;
+import net.fabricmc.api.EnvType;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.text.Style;
@@ -30,6 +31,13 @@ public class OnDamagedEffects extends PotionEffectProperty {
             return EventResult.pass();
         });
 
+        if(smartin.miapi.Environment.isClient()){
+            setupClient();
+        }
+    }
+
+    @net.fabricmc.api.Environment(EnvType.CLIENT)
+    public void setupClient(){
         LoreProperty.loreSuppliers.add(itemStack -> {
             List<Text> lines = new ArrayList<>();
             for (EffectHolder effectHolder : getStatusEffects(itemStack)) {