+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 - added When Attacking, When Killing and onDeath Potion effects
 - added Damage Bonus property against Smite , aquatic and Arachnic Mobs
 - added onKill Explosion property
+- updated config
+- added config option to customize glint coloring